+++ .ado/publish.yml
@@ -42,10 +42,9 @@ extends:
 
     stages:
       - stage: main
-        displayName:
         jobs:
           - job: build
-            displayName: Build and test
+            displayName: Build, test, pack
             cancelTimeoutInMinutes: 45 # to allow more time to collect CodeQL debug info.
 
             pool:
@@ -76,8 +75,6 @@ extends:
                   version: '${{ variables.nodeVersion }}.x'
                   checkLatest: false
 
-              - script: npm i -g npm@${{ variables.npmVersion }}
-
               # For multiline scripts, we want the whole task to fail if any line of the script fails.
               # ADO doesn't have bash configured this way by default. To fix we override the SHELLOPTS built-in variable.
               # https://www.gnu.org/software/bash/manual/html_node/The-Set-Builtin.html
@@ -95,11 +92,13 @@ extends:
                 displayName: Show environment
 
               - script: yarn --frozen-lockfile
+                displayName: Install dependencies
 
-              - script: yarn build
-              - script: yarn test:unit
-              - script: yarn test:func
-              - script: yarn test:e2e
+              - script: yarn build --verbose
+                displayName: Build
+
+              - script: yarn test --verbose
+                displayName: Test
 
               # TODO: figure out if this should bump or not (currently does bump)
               - script: |
