+++ .ado/release.yml
@@ -15,13 +15,13 @@ resources:
             - main
 
   repositories:
-    - repository: CustomPipelineTemplates
+    - repository: OfficePipelineTemplates
       type: git
       name: 1ESPipelineTemplates/OfficePipelineTemplates
       ref: refs/tags/release
 
 extends:
-  template: v1/Office.Official.PipelineTemplate.yml@CustomPipelineTemplates
+  template: v1/Office.Official.PipelineTemplate.yml@OfficePipelineTemplates
   parameters:
     pool:
       name: Azure-Pipelines-1ESPT-ExDShared
