+++ beachball.config.js
@@ -1,9 +1,10 @@
 // @ts-check
-/** @type {Partial<import('./src/types/BeachballOptions').RepoOptions>}*/
+/** @type {Partial<import('./packages/beachball/src/types/BeachballOptions').RepoOptions>}*/
 const config = {
   branch: 'main',
   commit: false,
   disallowedChangeTypes: ['major'],
+  groupChanges: true,
   ignorePatterns: ['.*ignore', '.eslintrc.js', 'jest.*.js', 'src/__*/**'],
 };
