+++ package.json
@@ -11,11 +11,12 @@
     "node": ">=14.0.0"
   },
   "workspaces": [
-    "packages/*"
+    "packages/*",
+    "scripts"
   ],
   "scripts": {
     "beachball": "node ./packages/beachball/lib/cli.js",
-    "build": "yarn workspace beachball build",
+    "build": "yarn lage build",
     "bump": "yarn beachball bump",
     "change": "yarn beachball change",
     "checkchange": "yarn beachball check",
@@ -24,19 +25,19 @@
     "format": "prettier --write .",
     "format:check": "prettier --check .",
     "prepare": "husky install",
-    "lint": "yarn lint:deps && yarn lint:marketplace && yarn lint:code",
-    "lint:code": "yarn workspace beachball lint",
-    "lint:deps": "yarn workspace beachball run depcheck .",
+    "lage": "cross-env NODE_OPTIONS=\"--experimental-abortcontroller\" lage",
+    "lint": "yarn lint:deps && yarn lint:versions && yarn lint:marketplace && yarn lint:code",
+    "lint:ci": "yarn lint:deps --verbose && yarn lint:versions && yarn lint:marketplace && yarn lint:code --verbose",
+    "lint:code": "yarn lage lint",
+    "lint:deps": "yarn lage depcheck",
+    "lint:versions": "yarn syncpack:check",
     "lint:marketplace": "node scripts/lint-marketplace.js",
     "release": "yarn beachball publish -y",
     "release:docs": "echo \"Run this from the docs folder instead\" && exit 1",
-    "test": "yarn workspace beachball test",
-    "test:all": "yarn test:unit && yarn test:func && yarn test:e2e",
-    "test:e2e": "yarn workspace beachball test:e2e",
-    "test:func": "yarn workspace beachball test:func",
-    "test:unit": "yarn workspace beachball test:unit",
-    "test:watch": "yarn workspace beachball test:watch",
-    "update-snapshots": "yarn workspace beachball update-snapshots"
+    "syncpack:check": "syncpack list-mismatches",
+    "syncpack:update": "syncpack fix-mismatches",
+    "test": "yarn lage test",
+    "update-snapshots": "yarn lage update-snapshots"
   },
   "lint-staged": {
     "*": [
@@ -50,14 +51,17 @@
     "@typescript-eslint/eslint-plugin": "^5.0.0",
     "@typescript-eslint/parser": "^5.0.0",
     "@typescript-eslint/utils": "^5.0.0",
+    "cross-env": "^7.0.0",
     "depcheck": "^1.4.7",
     "eslint": "^8.0.0",
     "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-etc": "^2.0.3",
     "husky": "^8.0.0",
     "jest": "^29.0.0",
+    "lage": "2.4.0",
     "lint-staged": "^12.0.0",
     "prettier": "~2.8.4",
+    "syncpack": "^9.0.0",
     "ts-jest": "29.2.6",
     "typescript": "~5.2.0"
   },
@@ -68,9 +72,11 @@
     "**/body-parser/qs": "^6.14.2",
     "**/express/qs": "^6.14.2",
     "**/@cypress/request/qs": "^6.14.2",
+    "**/syncpack/minimatch": "^6.2.3",
+    "**/syncpack/semver": "^7.7.4",
     "**/verdaccio/handlebars": "^4.7.9",
-    "**/verdaccio/minimatch": "^7.4.9",
     "**/verdaccio/js-yaml": "^4.1.0",
+    "**/verdaccio/minimatch": "^7.4.9",
     "**/verdaccio/validator": "^13.15.22",
     "**/@verdaccio/config/js-yaml": "^4.1.0",
     "**/@verdaccio/config/minimatch": "^7.4.9",
@@ -79,14 +85,16 @@
   },
   "rationale": {
     "devDependencies": {
+      "lage@2.4.0": "Pin to newest version that supports node 14",
       "ts-jest@29.2.6": "29.3.0 introduces type-fest v4 which requires node 16+"
     },
     "resolutions": {
       "**/lodash": "Unpin due to security issue",
       "**/*/qs": "Unpin due to security issue",
+      "**/semver": "Unpin due to security issues",
       "**/verdaccio/*": "Unpin due to security issues",
       "**/@verdaccio/*/*": "Unpin due to security issues",
-      "**/minimatch": "Unpin minimatch v7 due to security issues"
+      "**/minimatch": "Unpin minimatch due to security issues"
     }
   }
 }
