--- packages/beachball/.eslintignore
@@ -1,3 +1,0 @@
-scripts
-docs
-/*.config.js
