+++ scripts/config/eslintrc.base.js
@@ -1,0 +1,152 @@
+// @ts-check
+
+/** @type {import('@typescript-eslint/utils').TSESLint.Linter.Config} */
+const config = {
+  parser: '@typescript-eslint/parser',
+  parserOptions: {
+    ecmaVersion: 2020,
+  },
+  plugins: ['@typescript-eslint', 'etc'],
+  reportUnusedDisableDirectives: true,
+  extends: [
+    'eslint:recommended',
+    'plugin:@typescript-eslint/recommended',
+    'plugin:@typescript-eslint/recommended-requiring-type-checking',
+    'prettier',
+  ],
+  ignorePatterns: ['lib/**/*', 'bin/**/*', '.eslintrc.js', '*.config.js'],
+  rules: {
+    '@typescript-eslint/consistent-generic-constructors': 'error',
+    '@typescript-eslint/consistent-type-assertions': 'error',
+    '@typescript-eslint/consistent-type-imports': 'error',
+    '@typescript-eslint/consistent-type-exports': 'error',
+    '@typescript-eslint/explicit-module-boundary-types': 'error',
+    '@typescript-eslint/naming-convention': [
+      'error',
+      {
+        selector: 'variableLike',
+        format: ['PascalCase', 'camelCase'],
+        leadingUnderscore: 'allow',
+      },
+    ],
+    '@typescript-eslint/no-non-null-assertion': 'error',
+    '@typescript-eslint/no-shadow': ['error', { ignoreTypeValueShadow: true }],
+    '@typescript-eslint/no-unused-vars': [
+      'error',
+      {
+        args: 'after-used',
+        // Follow the typescript pattern of ignoring things starting with _
+        argsIgnorePattern: '^_',
+        destructuredArrayIgnorePattern: '^_',
+        varsIgnorePattern: '^_',
+        ignoreRestSiblings: true,
+      },
+    ],
+    '@typescript-eslint/prefer-for-of': 'error',
+    'etc/no-deprecated': 'error',
+    // ESLint has a built-in version of this rule without the allowTypeImports functionality.
+    // Some overrides use the built-in ESLint rule to set additional restricted imports without
+    // overriding these defaults.
+    '@typescript-eslint/no-restricted-imports': [
+      'error',
+      {
+        paths: [
+          {
+            name: 'node:test',
+            message: 'You probably meant to import from "@jest/globals"',
+          },
+        ],
+      },
+    ],
+    'no-restricted-properties': [
+      'error',
+      {
+        object: 'fs',
+        property: 'promises',
+        message: 'Import "fs/promises" directly instead. This allows methods to be mocked with Jest.',
+      },
+      {
+        object: 'process',
+        property: 'cwd',
+        message: 'Pass the proper cwd through to avoid accidentally running operations in the wrong context.',
+      },
+      {
+        object: 'process',
+        property: 'chdir',
+        message:
+          'beachball should not be dependent on the actual process.cwd(). Ensure the proper cwd is passed through instead.',
+      },
+      {
+        object: 'process',
+        property: 'exit',
+        message: 'Errors should be propagated to the top level and handled there.',
+      },
+    ],
+
+    // Downgrade these rules to warnings because they cause excessive/unhelpful noise when
+    // the actual problem is type errors due to a missing import...
+    // With the rules as warnings, there will only be "red squiggles" at the spot of the type error
+    // rather than sometimes covering the whole statement.
+    '@typescript-eslint/no-unsafe-argument': 'warn',
+    '@typescript-eslint/no-unsafe-assignment': 'warn',
+    '@typescript-eslint/no-unsafe-call': 'warn',
+    '@typescript-eslint/no-unsafe-member-access': 'warn',
+    '@typescript-eslint/no-unsafe-return': 'warn',
+
+    // Disabled rules from extended configs:
+    'no-undef': 'off',
+    'no-regex-spaces': 'off',
+    // This rule requires using Record instead of index signatures, and it's disabled because
+    // sometimes the key name can be useful for documentation.
+    '@typescript-eslint/consistent-indexed-object-style': 'off',
+    // Don't restrict types of `template expression ${operands}`.
+    '@typescript-eslint/restrict-template-expressions': 'off',
+    // Right now this rule gives unnecessary warnings about comparisons with string literals,
+    // which the compiler would flag anyway. It could be helpful if there was a future fix or
+    // option to ignore those types of safe comparisons.
+    '@typescript-eslint/no-unsafe-enum-comparison': 'off',
+  },
+  overrides: [
+    {
+      files: ['*.js', '*.mjs'],
+      parserOptions: {
+        sourceType: 'module',
+      },
+      rules: {
+        // Rule doesn't handle JS files
+        '@typescript-eslint/explicit-module-boundary-types': 'off',
+      },
+    },
+    {
+      files: ['**/__*/**', '**/*.test.*'],
+      rules: {
+        '@typescript-eslint/no-non-null-assertion': 'off',
+        '@typescript-eslint/no-empty-function': 'off',
+        'no-restricted-properties': [
+          'error',
+          ...['describe', 'it', 'test']
+            .map(func => [
+              { object: func, property: 'only', message: 'Do not commit .only() tests' },
+              { object: func, property: 'skip', message: 'Do not commit .skip() tests (disable this rule if needed)' },
+            ])
+            .flat(),
+        ],
+        // Use the ESLint version of the rule to avoid overriding the restricted imports from the base config
+        'no-restricted-imports': [
+          'error',
+          {
+            paths: [
+              {
+                name: '@jest/globals',
+                importNames: ['xdescribe', 'xit', 'xtest'],
+                message: 'Do not commit disabled tests (disable this rule if needed)',
+              },
+            ],
+          },
+        ],
+      },
+    },
+  ],
+};
+
+module.exports = config;
