+++ scripts/config/tsconfig.base.json
@@ -1,0 +1,20 @@
+{
+  "compilerOptions": {
+    // compatible with node 14
+    "target": "ES2020",
+    "module": "commonjs",
+    "moduleResolution": "node",
+    "declaration": true,
+    "declarationMap": true,
+    "sourceMap": true,
+    "strict": true,
+    "noUnusedLocals": true,
+    "forceConsistentCasingInFileNames": true,
+    "esModuleInterop": true,
+    "allowJs": true,
+    "checkJs": true,
+    "skipLibCheck": true,
+    "lib": ["ES2020"],
+    "types": ["node"]
+  }
+}
